requirements.txt
factory-boy
asgiref==3.8.1
astroid==2.5.3
attrs==22.2.0
certifi==2021.10.8
cfgv==3.4.0
charset-normalizer==2.0.7
cloudinary==1.42.2
colorama==0.4.4
distlib==0.4.0
Django==5.1.6
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-crispy-forms==1.11.2
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.9
filelock==3.18.0
identify==2.6.13
idna==3.3
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
isort==5.8.0
itypes==1.2.0
Jinja2==3.1.5
jsonschema==3.2.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.6.0
Markdown==3.7
MarkupSafe==2.0.1
mccabe==0.6.1
nodeenv==1.9.1
packaging==21.3
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pylint==2.7.4
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.0.1
python-http-client==3.3.4
pytz==2021.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.26.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.13.2
sendgrid==6.9.3
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
toml==0.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.33.1
wrapt==1.12.1
zipp==3.21.0
